RDKB-57439 : [XER10-UK] IPv6 Handling - Wan & Lan Management.
Reason for change:
IPv6 Handling Requirement for Sky Products are different from Comcast. Sky using IAPD , DHCPv6 prefix delegation to provide prefixes for use on the subscriber LAN as well as to get WAN IPv6 address from backend.

The implementation for this feature is under CcspPandM components so that needs to be enable for applicable products like Sky. So we need to control _RDKB_GLOBAL_PRODUCT_REQ_ common MACRO with feature applicable flags with respect to region.

Test Procedure:
1. IPv6 ULA should work for Sky but those params writable permission should be prohibited for comcast products.
2. WAN and LAN functionality should work without any issue.

Risks: Medium

Signed-off-by: Lakshminarayanan <lakshminarayanan.shenbagaraj2@sky.uk>